% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluationFinal.R
\name{evaluation.metric.final}
\alias{evaluation.metric.final}
\title{Evaluation function for the estimated chain graph}
\usage{
evaluation.metric.final(est_Omega, est_B, true_Omega, true_B)
}
\arguments{
\item{est_Omega}{The estimated Omega}

\item{est_B}{The estimated B}

\item{true_Omega}{The true Omega}

\item{true_B}{The true B}
}
\value{
A list of evaluation results: Recall, Precision, MCC and SHD
}
\description{
The function to evaluate the numerical performances of methods
in terms of the estimation accuracy of undirected edges, directed edges and
the overall chain graph.
}
\examples{
\donttest{
evaluation.metric.final(est_Omega=matrix(rbinom(100,1,0.5),10,10),
         est_B=diag(10), true_Omega=matrix(rbinom(100,1,0.5),10,10),
         true_B=diag(10))
}

}
