% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recover_CG.R
\name{Recover.CG}
\alias{Recover.CG}
\title{Function to reconstruct the chain graph}
\usage{
Recover.CG(X, lambda1_n, lambda2_n, kappa_n, nu_n)
}
\arguments{
\item{X}{The data matrix, n*p matrix}

\item{lambda1_n}{Regularization parameter corresponding to the sparsity of Omega}

\item{lambda2_n}{Regularization parameter corresponding to the low-rank of L}

\item{kappa_n}{A hard threshold for truncated singular value decomposition(SVD)}

\item{nu_n}{An element-wise hard threshold for estimated B}
}
\value{
The estimated Omega and B
}
\description{
The function to reconstruct the chain graph / estimate (Omega, B).
}
\examples{
gen_data = simul.CG.Ex1(p.node = 50, n.sample = 500, p.un = 0.03,
                       p.hub = 0.2, p.hub.dir = 0.8, CG.seed = 1)
X = gen_data$X
est_res = Recover.CG(X = X, lambda1_n = 0.4*500^(-3/8),
                    lambda2_n = 0.8*500^(-3/8), kappa_n = 5.5*500^(-3/8),
                    nu_n = 3.8*500^(-1/4))
est_Omega = est_res$est_Omega
est_B = est_res$est_B

}
