% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recover_B.r
\name{Recover_B.R}
\alias{Recover_B.R}
\title{Function to recover B}
\usage{
Recover_B.R(X, LayerOrder, kappa_n, nu_n)
}
\arguments{
\item{X}{The data matrix, n*p matrix}

\item{LayerOrder}{The causal ordering among chain components}

\item{kappa_n}{A hard threshold for truncated singular value decomposition(SVD)}

\item{nu_n}{An element-wise hard threshold for estimated B}
}
\value{
The estimated B, p*p matrix
}
\description{
The function to recover/estimate B.
}
