% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recover_Omega.r
\name{Recover_Omega.R}
\alias{Recover_Omega.R}
\title{Function to recover the Omega}
\usage{
Recover_Omega.R(Sigma, lambda1, lambda2, init, maxiter_inner, maxiter_outer, mu, rho,
                       delta, abs_tol, rel_tol,
                       print_progress_inner, print_every_inner,
                       print_progress_outer, print_every_outer)
}
\arguments{
\item{Sigma}{The sample covariance matrix}

\item{lambda1}{Regularization parameter corresponding to the sparsity of Omega}

\item{lambda2}{Regularization parameter corresponding to the low-rank of L}

\item{init}{The initial value for ADMM algorithm}

\item{maxiter_inner}{The maximum iteration of inner ADMM algorithm}

\item{maxiter_outer}{The maximum iteration of outer ADMM algorithm}

\item{mu}{Stepsize of the ADMM algorithm}

\item{rho}{Stepsize of the ADMM algorithm}

\item{delta}{Tolerance for updating Phi}

\item{abs_tol}{Absolute tolerance to stop the algorithm}

\item{rel_tol}{Relative tolerance to stop the algorithm}

\item{print_progress_inner}{Whether the inner ADMM algorithm should report its progress}

\item{print_every_inner}{How often should the inner ADMM algorithm report its progress}

\item{print_progress_outer}{Whether the outer ADMM algorithm should report its progress}

\item{print_every_outer}{How often should the outer ADMM algorithm report its progress}
}
\value{
A list including all parameter matrices, the value of objective function and history
}
\description{
The function to recover the Omega via ADMM algorithm.
}
