% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recover_order.R
\name{Recover_order.R}
\alias{Recover_order.R}
\title{Function to recover the causal ordering of chain components}
\usage{
Recover_order.R(X,Omega)
}
\arguments{
\item{X}{The data matrix, n*p matrix}

\item{Omega}{The precision matrix of noise Omega, p*p matrix}
}
\value{
The estimated casual ordering of chain components
}
\description{
The function to recover the causal ordering of chain components.
}
