% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_CGFinal.R
\name{simul.CG.Ex1}
\alias{simul.CG.Ex1}
\title{Function to generate random chain graph (Example 1)}
\usage{
simul.CG.Ex1(p.node, n.sample, p.un, p.hub, p.hub.dir, CG.seed)
}
\arguments{
\item{p.node}{Number of nodes}

\item{n.sample}{Sample size}

\item{p.un}{The probability of connecting two nodes with undirected edge}

\item{p.hub}{The probability of the number of hub nodes in each chain component}

\item{p.hub.dir}{The probability of directed edges pointing from a hub node to the nodes in the lower chain components}

\item{CG.seed}{The random number for generating random chain graph}
}
\value{
A list including
\describe{
\item{Omega}{The precision matrix of noise}
\item{B}{The coefficient matrix corresponding to the directed edges}
\item{X}{The data matrix}
}
}
\description{
The function to generate random chain graph.
}
\details{
Particularly, we randomly connect each pair of nodes by undirected
edge with a certain probability
and read off multiple chain components. Then, a causal ordering of the chain
components is randomly assigned. For each chain component, we randomly select
the nodes as hubs with a certain probability, and let each hub node points to
the nodes in lower chain components with a certain probability.
}
\examples{
simul.CG.Ex1(p.node = 10, n.sample = 200, p.un = 0.1, p.hub = 0.5, p.hub.dir = 0.5, CG.seed = 1)


}
